requirements.txt
alt-profanity-check==1.0.2.1
amqp==5.0.9
ansimarkup==1.5.0
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bjoern==3.1.0
boto3==1.20.34
botocore==1.23.34
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.5
Django==4.0.1
django-celery-beat==2.3.0
django-ckeditor==6.2.0
django-configurations==2.3.2
django-configurations-celery==2020.10.26
django-cors-headers==3.11.0
django-countries==7.2.1
django-extensions==3.1.5
django-filter==22.1
django-ipware==4.0.2
django-js-asset==1.2.2
django-phonenumber-field==6.0.0
django-postgres-extra==2.0.3
django-rest-framework-condition==0.1.1
django-sendgrid-v5==1.2.0
django-storages==1.12.3
django-timezone-field==5.0
djangorestframework==3.13.1
enforce==0.3.4
filelock==3.8.0
html-sanitizer==1.9.3
idna==3.3
jellyfish==0.9.0
jmespath==0.10.0
joblib==1.1.0
kombu==5.2.3
langdetect==1.0.9
lxml==4.7.1
Markdown==3.4.1
networkx==2.6.3
numpy==1.22.0
packaging==21.3
pdfminer.six==20211012
pdfplumber==0.6.0
phonenumbers==8.12.40
Pillow==9.0.0
platformdirs==2.5.2
prompt-toolkit==3.0.26
psycopg2==2.9.3
pycodestyle==2.9.0
pycparser==2.21
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python-http-client==3.3.7
pytz==2021.3
pyuca==1.2
PyYAML==6.0
redis==4.1.2
regex==2021.11.10
requests==2.27.1
s3transfer==0.5.0
scikit-learn==1.0.2
scipy==1.7.3
segtok==1.5.11
sendgrid==6.9.7
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
structlog==21.5.0
tabulate==0.8.9
threadpoolctl==3.0.0
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.16.3
Wand==0.6.7
wcwidth==0.2.5
wrapt==1.13.3
yake==0.4.8
